Refactor to_chars
